requirements.txt
Django
Pillow
psycopg2
tzdata
python-dotenv
djangorestframework
asgiref
pyjwt
djangorestframework-simplejwt
django-filter
coverage
